refactor: Store 관련 DTO 위치 변경 및 기존 테스트 삭제
