+++ src/main/java/com/burntoburn/easyshift/dto/store/StoreUserDTO.java
@@ -1,4 +1,4 @@
-package com.burntoburn.easyshift.dto.store.res;
+package com.burntoburn.easyshift.dto.store;
 
 import com.burntoburn.easyshift.dto.user.UserDTO;
 import lombok.Builder;
