+++ src/main/java/com/burntoburn/easyshift/service/StoreService.java
@@ -7,10 +7,10 @@ import com.burntoburn.easyshift.dto.shift.res.AssignedShiftDTO;
 import com.burntoburn.easyshift.dto.shift.res.ShiftDateDTO;
 import com.burntoburn.easyshift.dto.shift.res.ShiftGroupDTO;
 import com.burntoburn.easyshift.dto.shift.res.ShiftKey;
-import com.burntoburn.easyshift.dto.store.req.StoreCreateRequest;
-import com.burntoburn.easyshift.dto.store.res.StoreDto;
-import com.burntoburn.easyshift.dto.store.res.StoreScheduleResponseDTO;
-import com.burntoburn.easyshift.dto.store.res.StoreUserDTO;
+import com.burntoburn.easyshift.dto.store.StoreCreateRequest;
+import com.burntoburn.easyshift.dto.store.StoreDto;
+import com.burntoburn.easyshift.dto.store.StoreScheduleResponseDTO;
+import com.burntoburn.easyshift.dto.store.StoreUserDTO;
 import com.burntoburn.easyshift.dto.user.UserDTO;
 import com.burntoburn.easyshift.entity.schedule.Schedule;
 import com.burntoburn.easyshift.entity.schedule.Shift;
